little better

+++ src/src/main/java/org/openjfx/model/Location.java
@@ -1,27 +1,27 @@
 package org.openjfx.model;
 
 public class Location {
-    private int positionX;
-    private int positionY;
+    private double positionX;
+    private double positionY;
 
-    public Location(int positionX, int positionY) {
+    public Location(double positionX, double positionY) {
         this.positionX = positionX;
         this.positionY = positionY;
     }
 
-    public int getPositionX() {
+    public double getPositionX() {
         return positionX;
     }
 
-    public void setPositionX(int positionX) {
+    public void setPositionX(double positionX) {
         this.positionX = positionX;
     }
 
-    public int getPositionY() {
+    public double getPositionY() {
         return positionY;
     }
 
-    public void setPositionY(int positionY) {
+    public void setPositionY(double positionY) {
         this.positionY = positionY;
     }
 }
